marketplace_server/pyproject.toml
[tool.poetry]
name = "marketplace-server"
version = "0.1.0"
description = "Django, skillbox, backend server for marketplace"
authors = ["Дмитрий Бондарев <baev@cumproject.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8, <=3.13"
django = "4.2.7"
ruff = "^0.1.4"
psycopg = "^3.1.12"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
django-allauth = "^0.58.2"
django-extensions = "^3.2.3"
celery = "^5.3.6"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
redis = "^5.0.1"
croniter = "^2.0.1"

[tool.ruff]
line-length = 120
exclude = [
    "migrations",
    "manage.py",
]

[tool.ruff.lint]
unfixable = ["B"]
select = [
    "F",
    "B",
    "E",
    "I",
    "UP",
    "N",
    "ANN",
    "Q"
]
ignore = ["ANN002", "ANN003", "ANN101", "ANN102", "ANN204"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
